Create P2P Encrypted Chat
Encrypted chat with manual connection to the counterparty. The encryption is so good that not only third parties have no access to the correspondence, but also the people themselves participating in this correspondence :) but in general, if you remove or block the encryption, it works perfectly (even if this violates the main idea).
Basic commands:
/connect 192.168.1.100:52345 - Connect to another node
> /peers - Show active connections
> /exit - Exit the program
For this to work (works exclusively on a local network) on the second PC you need to replace BootstrapPeer with the current IP of the first and also assign instead of 30000 the UDP port that will be written to you after running the command.